App and drawing works! Just needs SD
So much stuff. Got rid of status.h because it was pretty useless with pet.h next to it. renamed countdown to tick_stats to avoid name collisions, and used the millisecond timer for tick_stats rather than systick. Lots of system calls added and changed. Probably a bunch of other stuff too. pet.c was added and pet.h changed a lot. Among other things.